Move portlayer/vsphere/storage into storage/vsphere pkg
I wanted to keep all the vsphere specific implementation stuff for all
of our portlayer handlers in the vsphere package but turns out no other
module was doing this which led to some confusion.  Might was well keep
this vsphere specific storage implementation in the portlayer storage
package.
