+++ portlayer/storage/vsphere/store.go
