+++ portlayer/storage/vsphere/store_test.go
